refactor: change directories for better discernment
