+++ src/infra/adapter/input/cli/video-input.ts
