+++ src/infra/adapter/output/external-services/frame-extractor-ffmpeg.ts
